AzureFunctions/requirements.txt
azure-functions
azure-data-tables
azure-identity
selenium
requests
pandas
python-dateutil
beautifulsoup4
lxml
